+++ lib/src/model/snacky.dart
@@ -2,6 +2,7 @@ import 'package:flutter/widgets.dart';
 import 'package:snacky/src/model/cancelable_snacky.dart';
 import 'package:snacky/src/model/snacky_location.dart';
 import 'package:snacky/src/model/snacky_type.dart';
+import 'package:snacky/src/util/duration/snacky_duration_util.dart';
 
 class Snacky {
   final String title;
@@ -13,12 +14,18 @@ class Snacky {
   final bool openUntillClosed;
   final bool canBeClosed;
   final VoidCallback? onTap;
-  final Duration showDuration;
+  final Duration? _showDuration;
   final Duration transitionDuration;
   final Curve transitionCurve;
   final SnackyLocation? location;
 
   final Widget Function(BuildContext, CancelableSnacky)? builder;
+
+  /// If null the `showDuration` will be calculated based on the length of the title and subtitle
+  Duration get showDuration {
+    if (_showDuration != null) return _showDuration;
+    return SnackyDurationUtil.calculateDuration(title, subtitle);
+  }
 
   const Snacky({
     required this.title,
@@ -30,19 +37,21 @@ class Snacky {
     this.onTap,
     this.canBeClosed = false,
     this.openUntillClosed = false,
-    this.showDuration = const Duration(seconds: 4),
+    Duration? showDuration,
     this.transitionDuration = const Duration(milliseconds: 250),
     this.transitionCurve = Curves.easeInOut,
     this.location,
-  }) : builder = null;
+  })  : builder = null,
+        _showDuration = showDuration;
 
   const Snacky.widget({
     required this.builder,
-    this.showDuration = const Duration(seconds: 4),
+    Duration? showDuration,
     this.transitionDuration = const Duration(milliseconds: 250),
     this.transitionCurve = Curves.easeInOut,
     this.location,
   })  : title = '',
+        _showDuration = showDuration,
         type = SnackyType.info,
         subtitle = null,
         leadingWidgetBuilder = null,